refactor(ai): Stop using useEffect, build context inside the evaluate handler and rename the custom hook that manages our state!
