+++ apps/web/src/serlo-editor-integration/serlo-plugin-wrappers/exercise-serlo-static-renderer.tsx
@@ -32,7 +32,6 @@ export function ExerciseSerloStaticRenderer(props: EditorExerciseDocument) {
   const isRevisionView = useContext(RevisionViewContext)
 
   const exerciseContextValue = useBuildExerciseContext(props)
-  console.log('Building exercise context value: ', exerciseContextValue)
 
   const context = props.serloContext
